Add test for storage providers
