+++ pkg/component/etcd/statefulset/statefulset_suite_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package statefulset
+package statefulset_test
 
 import (
 	"testing"
